+++ scripts/save.sh
@@ -39,6 +39,8 @@ pane_format() {
 	format+="${delimiter}"
 	format+="#{pane_index}"
 	format+="${delimiter}"
+	format+="#{pane_title}"
+	format+="${delimiter}"
 	format+=":#{pane_current_path}"
 	format+="${delimiter}"
 	format+="#{pane_active}"
@@ -142,46 +144,6 @@ capture_pane_contents() {
 	fi
 }
 
-save_shell_history() {
-	if [ "$pane_command" = "bash" ]; then
-		local history_w='history -w'
-		local history_r='history -r'
-		local accept_line='C-m'
-		local end_of_line='C-e'
-		local backward_kill_line='C-u'
-	elif [ "$pane_command" = "zsh" ]; then
-		# fc -W does not work with -L
-		# fc -l format is different from what's written by fc -W
-		# fc -R either reads the format produced by fc -W or considers
-		# the entire line to be a command. That's why we need -n.
-		# fc -l only list the last 16 items by default, I think 64 is more reasonable.
-		local history_w='fc -lLn -64 >'
-		local history_r='fc -R'
-
-		local zsh_bindkey="$(zsh -i -c bindkey)"
-		local accept_line="$(expr "$(echo "$zsh_bindkey" | grep -m1 '\saccept-line$')" : '^"\(.*\)".*')"
-		local end_of_line="$(expr "$(echo "$zsh_bindkey" | grep -m1 '\send-of-line$')" : '^"\(.*\)".*')"
-		local backward_kill_line="$(expr "$(echo "$zsh_bindkey" | grep -m1 '\sbackward-kill-line$')" : '^"\(.*\)".*')"
-	else
-		return
-	fi
-
-	local pane_id="$1"
-	local pane_command="$2"
-	local full_command="$3"
-	if [ "$full_command" = ":" ]; then
-		# leading space prevents the command from being saved to history
-		# (assuming default HISTCONTROL settings)
-		local write_command=" $history_w '$(resurrect_history_file "$pane_id" "$pane_command")'"
-		local read_command=" $history_r '$(resurrect_history_file "$pane_id" "$pane_command")'"
-		# C-e C-u is a Bash shortcut sequence to clear whole line. It is necessary to
-		# delete any pending input so it does not interfere with our history command.
-		tmux send-keys -t "$pane_id" "$end_of_line" "$backward_kill_line" "$write_command" "$accept_line"
-		# Immediately restore after saving
-		tmux send-keys -t "$pane_id" "$end_of_line" "$backward_kill_line" "$read_command" "$accept_line"
-	fi
-}
-
 get_active_window_index() {
 	local session_name="$1"
 	tmux list-windows -t "$session_name" -F "#{window_flags} #{window_index}" |
@@ -227,14 +189,14 @@ fetch_and_dump_grouped_sessions(){
 dump_panes() {
 	local full_command
 	dump_panes_raw |
-		while IFS=$d read line_type session_name window_number window_active window_flags pane_index dir pane_active pane_command pane_pid history_size; do
+		while IFS=$d read line_type session_name window_number window_active window_flags pane_index pane_title dir pane_active pane_command pane_pid history_size; do
 			# not saving panes from grouped sessions
 			if is_session_grouped "$session_name"; then
 				continue
 			fi
 			full_command="$(pane_full_command $pane_pid)"
 			dir=$(echo $dir | sed 's/ /\\ /') # escape all spaces in directory path
-			echo "${line_type}${d}${session_name}${d}${window_number}${d}${window_active}${d}${window_flags}${d}${pane_index}${d}${dir}${d}${pane_active}${d}${pane_command}${d}:${full_command}"
+			echo "${line_type}${d}${session_name}${d}${window_number}${d}${window_active}${d}${window_flags}${d}${pane_index}${d}${pane_title}${d}${dir}${d}${pane_active}${d}${pane_command}${d}:${full_command}"
 		done
 }
 
@@ -259,15 +221,8 @@ dump_state() {
 dump_pane_contents() {
 	local pane_contents_area="$(get_tmux_option "$pane_contents_area_option" "$default_pane_contents_area")"
 	dump_panes_raw |
-		while IFS=$d read line_type session_name window_number window_active window_flags pane_index dir pane_active pane_command pane_pid history_size; do
+		while IFS=$d read line_type session_name window_number window_active window_flags pane_index pane_title dir pane_active pane_command pane_pid history_size; do
 			capture_pane_contents "${session_name}:${window_number}.${pane_index}" "$history_size" "$pane_contents_area"
-		done
-}
-
-dump_shell_history() {
-	dump_panes |
-		while IFS=$d read line_type session_name window_number window_active window_flags pane_index dir pane_active pane_command full_command; do
-			save_shell_history "$session_name:$window_number.$pane_index" "$pane_command" "$full_command"
 		done
 }
 
@@ -299,9 +254,6 @@ save_all() {
 		dump_pane_contents
 		pane_contents_create_archive
 		rm "$(pane_contents_dir "save")"/*
-	fi
-	if save_shell_history_option_on; then
-		dump_shell_history
 	fi
 	remove_old_backups
 	execute_hook "post-save-all"