+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # Changelog
 
 ### master
+- Remove deprecated "restoring shell history" feature.
+
+### v4.0.0, 2022-04-10
 - Proper handling of `automatic-rename` window option.
+- save and restore tmux pane title (breaking change: you have to re-save to be
+  able to properly restore!)
 
 ### v3.0.0, 2021-08-30
 - save and restore tmux pane contents (@laomaiweng)